% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ils.fqcs.R
\name{boot.sim.set}
\alias{boot.sim.set}
\title{Bootstrap samples of a functional statistic}
\usage{
boot.sim.set(x, smo = 0.05)
}
\arguments{
\item{x}{An object of classs \code{fdata}.}

\item{smo}{The smoothing parameter for the bootstrap samples.}
}
\description{
\code{data.bootstrap} provides bootstrap samples for functional data.
}
