% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ils.fqcs.R
\name{outliers.ils}
\alias{outliers.ils}
\title{Detecting outliers for functional dataset}
\usage{
outliers.ils(x, fdep = depth.FM, trim = 0.01)
}
\arguments{
\item{x}{A \code{fdata}}

\item{fdep}{Type of depth measure, by default depth.mode.}

\item{trim}{The percentaje of the trimming, by default is 1\%}
}
\description{
Procedure for detecting funcitonal outliers.
}
