% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.lab.qcdata.R
\name{plot.lab.qcdata}
\alias{plot.lab.qcdata}
\title{Plot method for 'lab.qcdata' objects}
\usage{
\method{plot}{lab.qcdata}(x, xlab = NULL, ylab = NULL, col = "blue", ...)
}
\arguments{
\item{x}{An object of class \code{lab.qcdata} (Univariate Quality Control Data).}

\item{xlab}{Title for the x axis.}

\item{ylab}{Title for the y axis.}

\item{col}{Color of type material, when there only one.}

\item{...}{Other arguments to be passed to or from methods.}
}
\description{
Generic function for plotting objects of 'lab.qcdata' class. Results of  univariate ILS studies are graphically shown.
}
\examples{
library(ILS)
data(Glucose)
Glucose.qcdata <- lab.qcdata(Glucose)
str(Glucose.qcdata)
plot(Glucose.qcdata)
}
