% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.lab.qcs.R
\name{plot.lab.qcs}
\alias{plot.lab.qcs}
\title{Plot method for 'lab.qcs' objects}
\usage{
\method{plot}{lab.qcs}(x, title = NULL, xlab = NULL, ylab = NULL,
  col = NULL, ylim = NULL, ...)
}
\arguments{
\item{x}{Functional data object or a list with functional objects}

\item{title}{Main title for the plot}

\item{xlab}{Title for the x axis}

\item{ylab}{Title for the y axis}

\item{col}{Color specifications}

\item{ylim}{Numeric vectors of length 2 (coordinates ranges)}

\item{...}{arguments to be passed to or from methods.}
}
\description{
Generic function for plotting objects of 'lab.qcs' class. Results of  univariate ILS studies are graphically shown.
}
